Refactored TextField
