+++ src/Assets/Elements/StyledTextField/index.js
@@ -1,37 +1,24 @@
-import styled from 'styled-components';
-import TextField from '@material-ui/core/TextField';
+import React from 'react';
+import StyledTextField from './Style';
+import { useSelector } from 'react-redux';
 
-const StyledTextField = styled(TextField)`
-  width: ${(props) => (props.fullWidth ? null : '70%')};
-  background-color: #f8faff;
-  border-radius: 10px;
-  margin: 24px 0 24px 0;
-  transform-origin: top right;
-  .MuiInputBase-root {
-    input::-webkit-outer-spin-button,
-    input::-webkit-inner-spin-button {
-      -webkit-appearance: none;
-      margin: 0;
-    }
-  }
-  .MuiInputLabel-formControl {
-    right: 0;
-    left: unset;
-  }
-  .MuiInputLabel-shrink {
-    transform-origin: top right;
-    opacity: 0.6;
-    color: #1e2132;
-  }
-  .MuiSvgIcon-root {
-    color: ${(props) => props.iconColor};
-  }
-  .MuiInputBase-root {
-    color: #000b40;
-  }
-  .MuiFormHelperText-root {
-    text-align: right;
-  }
-`;
+/**
+ * @author Idan Gigi idangi@matrix.co.il
+ * @param {String} iconColor
+ * @param {String} width
+ * @param {MuiTextFieldProps}
+ * @returns {ReactElement}
+ */
+const CustomizedTextField = ({ width, iconColor, ...MuiTextFieldProps }) => {
+  const direction = useSelector((state) => state.settings.lang_dir);
+  return (
+    <StyledTextField
+      direction={direction}
+      {...MuiTextFieldProps}
+      width={width}
+      iconColor={iconColor}
+    />
+  );
+};
 
-export default StyledTextField;
+export default CustomizedTextField;
